chore(lib_): add printf_() agenda

+++ lib_/strlen_.c
@@ -4,7 +4,7 @@
  * @param s string
  * @return length of string
  */
-int strlen_(char *s) {
+size_t strlen_(char *s) {
   size_t i = 0;
   while (s[i] != '\0') {
     i++;
